chore: Upgrade ESLint to v9
