--- .eslintignore
@@ -1,26 +1,0 @@
-# common
-.github/renovate.json
-dist/
-esm/
-build/
-fixtures/
-byline.ts
-prism.ts
-charm.ts
-pnpm-lock.yaml
-generated-dmmf.ts
-
-# client
-packages/client/generator-build/
-packages/client/declaration/
-packages/client/runtime/
-packages/client/src/__tests__/types/
-packages/client/scripts/default-index.js
-
-# cli
-packages/cli/prisma-client/
-packages/cli/install/
-packages/cli/preinstall/
-packages/cli/**/tmp-*
-
-sandbox/
