+++ package.json
@@ -19,7 +19,7 @@
     "preinstall": "node ./scripts/only-allow-pnpm.js pnpm",
     "precommit": "lint-staged",
     "eslint": "eslint",
-    "lint": "eslint --fix --ext .ts .",
+    "lint": "eslint --fix .",
     "check-engines-override": "tsx scripts/check-engines-override.ts",
     "format": "prettier --write .",
     "prettier-check": "prettier --check .",
@@ -30,6 +30,9 @@
     "bundle-size": "pnpm i && pnpm -r dev && pnpm -r --parallel exec pnpm pack && cd packages/bundle-size && pnpm run create-gzip-files"
   },
   "devDependencies": {
+    "@eslint/compat": "1.0.1",
+    "@eslint/eslintrc": "3.1.0",
+    "@eslint/js": "9.3.0",
     "@microsoft/api-extractor": "7.43.8",
     "@prisma/engines": "workspace:*",
     "@sindresorhus/slugify": "1.1.2",
@@ -53,7 +56,7 @@
     "dotenv-cli": "7.4.2",
     "esbuild": "0.21.3",
     "esbuild-register": "3.5.0",
-    "eslint": "8.57.0",
+    "eslint": "9.3.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-eslint-comments": "3.2.0",
     "eslint-plugin-import": "2.29.1",
@@ -64,6 +67,7 @@
     "execa": "5.1.1",
     "format-util": "1.0.5",
     "fs-extra": "11.1.1",
+    "globals": "15.3.0",
     "globby": "11.1.0",
     "graphviz-mit": "0.0.9",
     "husky": "9.0.11",
