+++ packages/client/tests/e2e/prisma-client-imports/src/default.ts
@@ -1,4 +1,3 @@
-/* eslint-disable import/no-duplicates */
 import { D1Database } from '@cloudflare/workers-types'
 import { neon, neonConfig, Pool as NeonPool } from '@neondatabase/serverless'
 import { Client as PlanetScaleClient } from '@planetscale/database'
@@ -77,4 +76,3 @@ export const planetScalePrismaClient = new PrismaClient({
   adapter: new PrismaPlanetScale(planetScaleClient),
 })
 void planetScalePrismaClient.user.findMany()
-
